# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Epstein Files Public API",
    "version": "1.0.0",
    "description": "Public JSON API for dataset stats, search, and file-level lookup.\n\n[Back to Website](/)"
  },
  "externalDocs": {
    "description": "Back to Website",
    "url": "/"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/api/stats": {
      "get": {
        "summary": "Get aggregate stats and dataset list",
        "responses": {
          "200": {
            "description": "Stats payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/search": {
      "get": {
        "summary": "Search files with filters",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "EFTA ID prefix."
          },
          {
            "name": "dataset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Dataset filter (for example: 9)."
          },
          {
            "name": "altered",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["all", "yes", "no"],
              "default": "all"
            }
          },
          {
            "name": "hidden",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["all", "yes", "no"],
              "default": "all"
            }
          },
          {
            "name": "deleted",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["all", "yes", "no"],
              "default": "all"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search result page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/file/{eftaId}": {
      "get": {
        "summary": "Get a file by EFTA ID",
        "parameters": [
          {
            "name": "eftaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "EFTA01244748"
            },
            "description": "EFTA identifier (for example: EFTA01244748).",
            "example": "EFTA01244748"
          }
        ],
        "responses": {
          "200": {
            "description": "File detail payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileDetailResponse"
                },
                "examples": {
                  "defaultLookup": {
                    "summary": "Default EFTA lookup",
                    "value": {
                      "file": {
                        "efta_id": "EFTA01244748",
                        "parent_efta_id": null,
                        "dataset": "9",
                        "file_path": "IMAGES/EFTA01244748.pdf",
                        "page_count": 1,
                        "hidden": false,
                        "deleted": true,
                        "altered": false,
                        "shows_in_search": true,
                        "last_checked": "2026-02-24T00:00:00.000Z",
                        "doj_website_page": null,
                        "notes": null,
                        "original_hash": null,
                        "current_hash": null,
                        "sources": {
                          "original_link": "https://doj-files.geeken.dev/doj_zips/VOL00009/IMAGES/EFTA01244748.pdf",
                          "doj_link": "https://www.justice.gov/epstein/files/DataSet%209/EFTA01244748.pdf"
                        }
                      },
                      "parent": null,
                      "children": []
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Not found"
                    }
                  },
                  "required": ["error"]
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DatasetCount": {
        "type": "object",
        "properties": {
          "dataset": {
            "type": ["string", "null"]
          },
          "count": {
            "type": "integer"
          }
        },
        "required": ["dataset", "count"]
      },
      "Stats": {
        "type": "object",
        "properties": {
          "total_files": {
            "type": "integer"
          },
          "altered_files": {
            "type": "integer"
          },
          "hidden_files": {
            "type": "integer"
          },
          "deleted_files": {
            "type": "integer"
          },
          "files_with_parent": {
            "type": "integer"
          },
          "files_without_parent": {
            "type": "integer"
          },
          "distinct_datasets": {
            "type": "integer"
          },
          "avg_page_count": {
            "type": ["number", "null"]
          },
          "max_page_count": {
            "type": ["number", "null"]
          },
          "by_dataset": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DatasetCount"
            }
          }
        },
        "required": [
          "total_files",
          "altered_files",
          "hidden_files",
          "deleted_files",
          "files_with_parent",
          "files_without_parent",
          "distinct_datasets",
          "avg_page_count",
          "max_page_count",
          "by_dataset"
        ]
      },
      "FileRow": {
        "type": "object",
        "properties": {
          "efta_id": {
            "type": "string"
          },
          "parent_efta_id": {
            "type": ["string", "null"]
          },
          "dataset": {
            "type": ["string", "null"]
          },
          "file_path": {
            "type": ["string", "null"]
          },
          "page_count": {
            "type": ["integer", "null"]
          },
          "hidden": {
            "type": "boolean"
          },
          "deleted": {
            "type": "boolean"
          },
          "altered": {
            "type": "boolean"
          },
          "shows_in_search": {
            "type": ["boolean", "null"]
          },
          "last_checked": {
            "type": ["string", "null"]
          },
          "doj_website_page": {
            "type": ["integer", "null"]
          },
          "notes": {
            "type": ["string", "null"]
          },
          "original_hash": {
            "type": ["string", "null"]
          },
          "current_hash": {
            "type": ["string", "null"]
          },
          "sources": {
            "type": "object",
            "properties": {
              "original_link": {
                "type": ["string", "null"]
              },
              "doj_link": {
                "type": ["string", "null"]
              }
            },
            "required": ["original_link", "doj_link"]
          }
        },
        "required": [
          "efta_id",
          "parent_efta_id",
          "dataset",
          "file_path",
          "page_count",
          "hidden",
          "deleted",
          "altered",
          "shows_in_search",
          "last_checked",
          "doj_website_page",
          "notes",
          "original_hash",
          "current_hash",
          "sources"
        ]
      },
      "StatsResponse": {
        "type": "object",
        "properties": {
          "stats": {
            "$ref": "#/components/schemas/Stats"
          },
          "datasets": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["stats", "datasets"]
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileRow"
            }
          }
        },
        "required": ["total", "page", "pageSize", "totalPages", "rows"]
      },
      "FileDetailResponse": {
        "type": "object",
        "properties": {
          "file": {
            "$ref": "#/components/schemas/FileRow"
          },
          "parent": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/FileRow"
              },
              {
                "type": "null"
              }
            ]
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileRow"
            }
          }
        },
        "required": ["file", "parent", "children"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "string"
          }
        },
        "required": ["error"]
      }
    }
  }
}